オブジェクト削除

+++ jpl/src/ch20/ex01/TranslateByte.java
@@ -25,6 +25,6 @@ public class TranslateByte {
 	{
 		int b;
 		while ((b = in.read()) != -1)
-			out.write(b == from ? to : b);
+			out.write((byte) b == from ? to : (byte) b);
 	}
 }
